requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-ckeditor==6.4.2
django-etc==1.3.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
future==0.18.2
Pillow==9.2.0
pymongo==3.12.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.2.4
